renaming gpu implementation
