+++ src/xmipp/libraries/reconstruction_cuda/cuda_rot_polar_estimator.h
@@ -28,11 +28,12 @@
 
 #include "reconstruction/arotation_estimator.cpp"
 #include "reconstruction_cuda/gpu.h"
-#include "reconstruction_cuda/cuda_find_extrema.h"
 #include "cuda_fft.h"
 #include <thread>
 #include <condition_variable>
 #include <algorithm>
+
+#include "cuda_single_extrema_finder.h"
 
 namespace Alignment {
